(PC-18008) feat(marketing): rename _marketingAndCommunication > marketingAndCommunication and move in features/internal
