+++ src/features/internal/marketingAndCommunication/atoms/LocationFilterChoice.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { LocationFilterChoice } from 'features/_marketingAndCommunication/atoms/LocationFilterChoice'
+import { LocationFilterChoice } from 'features/internal/marketingAndCommunication/atoms/LocationFilterChoice'
 import { LocationType } from 'features/search/enums'
 import { render, fireEvent } from 'tests/utils'
