+++ src/features/internal/marketingAndCommunication/components/DeeplinksHistory.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useEffect } from 'react'
 import { FlatList } from 'react-native'
 import styled from 'styled-components/native'
 
-import { DeeplinkItem } from 'features/_marketingAndCommunication/atoms/DeeplinkItem'
-import { GeneratedDeeplink } from 'features/_marketingAndCommunication/components/DeeplinksGeneratorForm'
+import { DeeplinkItem } from 'features/internal/marketingAndCommunication/atoms/DeeplinkItem'
+import { GeneratedDeeplink } from 'features/internal/marketingAndCommunication/components/DeeplinksGeneratorForm'
 import { Checkbox } from 'ui/components/inputs/Checkbox/Checkbox'
 import { useSnackBarContext } from 'ui/components/snackBar/SnackBarContext'
 import { TouchableOpacity } from 'ui/components/TouchableOpacity'
