+++ src/features/internal/marketingAndCommunication/components/DeeplinksResult.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import styled from 'styled-components/native'
 
-import { DeeplinkItem } from 'features/_marketingAndCommunication/atoms/DeeplinkItem'
-import { GeneratedDeeplink } from 'features/_marketingAndCommunication/components/DeeplinksGeneratorForm'
+import { DeeplinkItem } from 'features/internal/marketingAndCommunication/atoms/DeeplinkItem'
+import { GeneratedDeeplink } from 'features/internal/marketingAndCommunication/components/DeeplinksGeneratorForm'
 import { getSpacing, Typo } from 'ui/theme'
 
 interface Props {
