+++ src/features/internal/marketingAndCommunication/pages/DeeplinksGenerator.native.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { DeeplinksGenerator } from 'features/_marketingAndCommunication/pages/DeeplinksGenerator'
+import { DeeplinksGenerator } from 'features/internal/marketingAndCommunication/pages/DeeplinksGenerator'
 import { render } from 'tests/utils'
 
 describe('<DeeplinksGenerator />', () => {
