+++ src/features/internal/marketingAndCommunication/pages/UTMParameters.native.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { UTMParameters } from 'features/_marketingAndCommunication/pages/UTMParameters'
+import { UTMParameters } from 'features/internal/marketingAndCommunication/pages/UTMParameters'
 import { render } from 'tests/utils'
 
 describe('<UTMParameters />', () => {
